requirements.txt
appdirs==1.4.1
Django==1.10.5
djangorestframework==3.5.4
packaging==16.8
pyparsing==2.1.10
six==1.10.0
